requirements.txt
Flask==2.3.3
requests==2.31.0
debugpy # Required for debugging.
streamlit==1.28.0
pandas==2.2.1
google-auth==2.28.1
cryptography==42.0.5
notion-client==1.0.0
google-cloud-storage==2.14.0
